third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-aware-create-shadow-root-expected.txt

CONSOLE WARNING: line 14: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4668884095336448 for more details.
Tests for ShadowAware JS APIs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS root1 is not root2
PASS root1 is window.internals.oldestShadowRoot(host)
PASS root2 is window.internals.youngestShadowRoot(host)
PASS successfullyParsed is true

TEST COMPLETE
